requirements.txt
Flask==3.1.0
PyPDF2==3.0.1
pdfplumber==0.11.6
pymupdf==1.25.5
requests==2.32.3
python-dotenv==1.1.0
Flask-SQLAlchemy==3.1.1
pytest==7.4.0
pytest-flask==1.3.0
SQLAlchemy==2.0.36
Werkzeug==3.1.3
Jinja2==3.1.6
elasticsearch==8.11.0
google-generativeai==0.3.2 
google-genai==1.0.0
google-api-core==2.27.0
beautifulsoup4==4.12.3
extract-msg==0.55.0
Pillow==11.0.0
msal==1.28.0
APScheduler==3.10.4
pytz==2024.1